fix: resolve pylint issues — move tests to separate file, remove unused import
Moved TestEvaluateMetricExceptionHandling to its own file to keep
test_processor.py under the 1000-line pylint limit. Removed unused
pytest import.
